Add 0.0.0.0/0 route if defaultGateway is true
On Windows, the internet connection is marked as "down" on VPN
connections where defaultGateway is set to true.  A workaround is to add
an extra route 0.0.0.0/0.  Ideally, the Windows client would handle
this, but as discussed in the eduVPN signal group, this can take a
while.  This workaround pushes the 0.0.0.0/0 route if the gateway is set
so that Microsoft Office works on Windows.  This change has no averse
effects on iOS, MacOS or Android.

When the Windows client handles this problem by itself, and the version
containing that fix is widely deployed, this patch can be removed again.