also push 0.0.0.0/0 when default gateway
in some situations, we are not sure exactly when, some Windows clients
think they do not have an Internet connection when connecting to a VPN
with the default gateway set. This workaround always pushes the
0.0.0.0/0 route in addition to OpenVPN's redirect-gateway "def1" flag.

This change also makes it impossible to push additional routes when the
default gateway is set, it makes no sense to push more routes so this
gets rid of that. The only reason this was needed was before the
0.0.0.0/0 became the default when defaultGateway was set to true.