move OpenVPN user/group "detection"/config to node
until now the portal/controller decided which openvpn `user` and
`group` were being used depending on the OS the portal/controller
was running on. As it is possible to have node(s) running on
different OSes it makes no sense to have it there. This
commit allows configuring/override it in the node config if
necessary or running on an unsupported platform.

References: https://todo.sr.ht/~eduvpn/server/133